Added Data Structure Classes

+++ src/kv_utility/ValueDetail.java
@@ -5,8 +5,8 @@ package kv_utility;
  */
 public class ValueDetail {
     private String value = null;
-    private double unixTS = 0;
-    private double hashed_value = 0;
+    private long unixTS = 0;
+    private int hashed_value = 0;
 
 
     public String getValue() {
@@ -17,19 +17,19 @@ public class ValueDetail {
         this.value = value;
     }
 
-    public double getUnixTS() {
+    public long getUnixTS() {
         return unixTS;
     }
 
-    public void setUnixTS(double unixTS) {
+    public void setUnixTS(long unixTS) {
         this.unixTS = unixTS;
     }
 
-    public double getHashed_value() {
+    public int getHashed_value() {
         return hashed_value;
     }
 
-    public void setHashed_value(double hashed_value) {
+    public void setHashed_value(int hashed_value) {
         this.hashed_value = hashed_value;
     }
 }
